+++ src/api/control/grpc/server.rs
@@ -357,12 +357,7 @@ impl ControlApi for ControlApiService {
                     let mut response = GetResponse::new();
                     match result {
                         Ok(elements) => {
-                            response.set_elements(
-                                elements
-                                    .into_iter()
-                                    .map(|(id, value)| (id.to_string(), value))
-                                    .collect(),
-                            );
+                            response.set_elements(elements);
                         }
                         Err(e) => {
                             response.set_error(e.into());